+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Add `around_iteration` hook
+
 ## 0.3.0 (2023-05-20)
 
 - Allow a default retry backoff to be configured